changed table values from int to decimal

+++ Server/migrations/20170915171152_coinValueTable.js
@@ -1,13 +1,12 @@
-
 exports.up = function(knex, Promise) {
   return Promise.all([
     knex.schema.createTable('coinValue', function(table) {
       table.increments('id').primary();
       table.integer('user').references('id').inTable('users').notNull().onDelete('cascade');
       table.string('coin');
-      table.integer('price');
-      table.integer('quantity');
-      table.integer('total');
+      table.decimal('price');
+      table.decimal('quantity');
+      table.decimal('total');
 
     })
   ])
